+++ SWFEM_Frag/cfjacobyupdate.cpp
@@ -1,7 +1,7 @@
 #include "cfjacobyupdate.h"
 
 void CFJacobyUpdate::exec(Data& src, Data& dst) {
-	for(unsigned int i = 0;i < num;i++) {		
+	for(size_t i = 0;i < num;i++) {		
 		NodeData& dt_src = src[src_start + i*src_step];
 		NodeData& dt_dst = dst[dst_start + i*dst_step];		
 		dt_dst.u += dt_src.u;
