+++ SWFEM_Frag/cfjacobyupdaterecv.cpp
@@ -1,7 +1,7 @@
 #include "cfjacobyupdaterecv.h"
 
 void CFJacobyUpdateRecv::exec(Data& dst, Data& buf) {
-	for(unsigned int i = 0;i < num;i++) {		
+	for(size_t i = 0;i < num;i++) {		
 		const NodeData& dt_src = buf[i];
 		NodeData& dt_dst = dst[start + i*step];		
 		dt_dst.u += dt_src.u;
