+++ SWFEM_Frag/cfloadpoints.cpp
@@ -4,16 +4,16 @@
 
 using namespace std;
 
-void CFLoadPoints::exec(MeshFragment& mesh, const std::string& filename, unsigned int step, unsigned int size) {
+void CFLoadPoints::exec(MeshFragment& mesh, const std::string& filename, size_t step, size_t size) {
 	std::ifstream f_in(filename.c_str(), ios::in);
 	if(!f_in.is_open()) {
 		cerr << "Error: can't open " << filename << endl;
 		return;
 	}
-	unsigned int num;
+	size_t num;
 		f_in >> num;
 	mesh.nodes.resize(num);
-	for(unsigned int i = 0; i < num;i++) {
+	for(size_t i = 0; i < num;i++) {
 		MeshNode& node = mesh.nodes[i];
 		f_in >> node.id >> node.lambda >> node.phi >> node.depth;
 	}
