+++ SWFEM_Frag/datafragment.h
@@ -71,7 +71,7 @@ public:
 	}
 
 	virtual ~DataFragment() {}
-	virtual unsigned int getWeight() = 0;
+	virtual size_t getWeight() = 0;
 };
 
 typedef std::vector<DataFragment*> DataFragmentPtrArray;
