+++ SWFEM_Frag/ieventmanager.h
@@ -7,9 +7,9 @@
  */
 class IEventManager {
 public:
-	//virtual void registerEventCF(const unsigned int& event_id, EventCompFragment *cf) = 0;
-	virtual void registerEventCFs(const unsigned int& event_id, const EventCompFragmentPtrArray& cfs) = 0;
-	virtual void waitForEvent(const unsigned int& event_id) = 0;
+	//virtual void registerEventCF(const size_t& event_id, EventCompFragment *cf) = 0;
+	virtual void registerEventCFs(const size_t& event_id, const EventCompFragmentPtrArray& cfs) = 0;
+	virtual void waitForEvent(const size_t& event_id) = 0;
 
 	virtual ~IEventManager() {}
 };
